requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
paho-mqtt==2.0.0
python-multipart==0.0.9
jinja2==3.1.3 
modbus==3.9.2